refactored acounting_services to use a journal builder. updated some bugs with fetching data for create_batch js file logic. updated tests acounting service to be more accurate and fixed some bugs in it . commented out  some old deprecated signals. updated some tables that used floats to use decimal instead and updated the corresponding services that used those models to use decimal. updated costing service to use the new right costing_unit_price  from the inventory log and not the old base_unit_price .

+++ gipcco_project/inventory/models/operational.py
@@ -124,7 +124,7 @@ class InventoryLog(models.Model):
         related_name='inventory_logs',
         verbose_name=_("Company")
     )
-    quantity = models.FloatField(verbose_name=_("Quantity"))
+    quantity = models.DecimalField(max_digits=14, decimal_places=4, verbose_name=_("Quantity"))
     timestamp = models.DateTimeField(verbose_name=_("Timestamp"))
     qc_no = models.CharField(
         max_length=100,
